requirements.txt
azure-functions==1.3.0
azure-storage-blob==12.3.2
numpy==1.19.1
